Added the .travis.yml file to work with Travis
Travis (travis-ci.org) provides automatic build (and CHECK) of the package whenever we push it to our GitHub repository!

The .travis.yml file was added to .Rbuildignore to avoid a NOTE when checking the package.

Also added envnames.Rproj to the tracking tree as suggested by Colin Gillespi at the useR!2019 workshop @Toulouse.